Auth added

+++ app/Http/Controllers/BlogController.php
@@ -3,17 +3,27 @@
 namespace App\Http\Controllers;
 
 use App\Category;
+use App\User;
+use Gate;
 use Illuminate\Http\Request;
 use App\Post;
+use Illuminate\Support\Facades\Auth;
 
 class BlogController extends Controller
 {
+    public function __construct(){
+        $this->middleware('auth', ['except' => ['index', 'showFull']]);
+    }
+
     public function index(){
 
         $posts = Post::paginate(3);
         $cat = Category::all();
+        $users = User::all();
+        $auth = Auth::user();
+//        dd($auth);
 
-        return view('blog_theme.pages.home', compact('posts','cat'));
+        return view('blog_theme.pages.home', compact('posts','cat', 'users', 'auth'));
     }
     public function addPost(){
 
@@ -25,14 +35,21 @@ class BlogController extends Controller
         $validateData = $request->validate([
             'title' => 'required|unique:posts|max:255',
             'body' => 'required',
-            'category'=> 'required'
+            'category'=> 'required',
+            'image'=>'mimes:jpeg, jpg, png, gif|required|max:10000'
         ]);
-//        dd($request);
+
+        $path = $request->file('image')->store('public/images');
+        $filename = str_replace('public/', "", $path);
+//        dd(Auth::id());
         Post::create([
             'title' => request('title'),
+            'body' => request('body'),
             'category' => request('category'),
-            'body' => request('body')
+            'image' => $filename,
+            'user_id' => Auth::id()
         ]);
+
         return redirect('/');
     }
 
@@ -41,23 +58,33 @@ class BlogController extends Controller
     }
 
     public function edit(Post $post){
-        return view('blog_theme.pages.edit', compact('post'));
+        if(Gate::allows('update', $post))
+        {
+
+            return view('blog_theme.pages.edit', compact('post' ));
+        }
+        dd('klaida');
+
     }
     public function storeUpdate(Request $request, Post $post){
-            Post::where('id', $post->id)->update($request->only(['title', 'category', 'body']));
+            Post::where('id', $post->id)->update($request->only(['title', 'category', 'body',]));
             return redirect('/post/'.$post->id);
     }
 
     public function delete(Post $post){
-        $post->delete();
-        return redirect('/');
+        if(Gate::allows('update', $post))
+        {
+            $post->delete();
+            return redirect('/');
+        }
+        dd('klaida');
     }
 //----- neveikia sitas
-    public function showByCategory(Request $request, Post $post1){
+    public function showByCategory(Request $request){
 
-        $posts = Post::with('categoryBy')->where('category', $request->category)->get();
-        $post1 = Post::paginate(2);
+        $posts = Post::with('categoryBy')->where('category', $request->category)->paginate(2);
+        ;
        // dd($posts);
-        return view('blog_theme.pages.category-one', compact('posts', 'post1'));
+        return view('blog_theme.pages.category-one', compact('posts'));
     }
 }
