+++ resources/views/blog_theme/pages/category-one.blade.php
@@ -23,20 +23,28 @@
                             <p class="post-meta mb-0">Category: {{$post->category}}</p>
                             <p class="post-meta mt-0">Created at: {{$post->created_at}}</p>
 
-                            <ul class="list-group">
-                                <li class="list-group-item"><a href="/edit/{{$post->id}}">Edit</a></li>
-                                <li class="list-group-item"><a href="/delete/{{$post->id}}">Delete</a></li>
-                            </ul>
+                    @foreach($users as $user)
+                        @if($user->id === $post->user_id)
+                            <p class="post-meta mt-0"> Author: {{ $user->name }}</p>
+                            @if(!empty($auth) and $auth->name === $user->name)
+                                <ul class="list-group">
+                                    <li class="list-group-item"><a href="/edit/{{$post->id}}">Edit</a></li>
+                                    <li class="list-group-item"><a href="/delete/{{$post->id}}">Delete</a></li>
+                                </ul>
+                            @endif
+                        @endif
+
+                    @endforeach
                 </div>
                 <hr>
             @endforeach
 
 
         <!-- Pager -->
-            <div class="clearfix">
+            <div class="clearfix ">
 {{--                 gauname nauja postu masyva, todel netinka sitas metodas--}}
-{{--                {{$post1->links()}}--}}
-                <a class="btn btn-primary float-left" href="/add-category">Back &rarr;</a>
+                {{$posts->links()}}
+                <a class="btn btn-secondary float-left" href="/add-category">Back &larr;</a>
             </div>
         </div>
     </div>
